+++ src/entities/script.cpp
@@ -14,6 +14,7 @@
 
 #include "script.h"
 
+#include <libraries/md4c/src/md4c-html.h>
 #include <libraries/versionnumber/versionnumber.h>
 #include <services/cryptoservice.h>
 #include <services/metricsservice.h>
@@ -24,6 +25,7 @@
 #include <QDebug>
 #include <QDir>
 #include <QJsonDocument>
+#include <QRegularExpression>
 #include <QSqlDatabase>
 #include <QSqlError>
 #include <QSqlQuery>
@@ -31,7 +33,13 @@
 #include <QtCore/QJsonArray>
 
 const QString Script::ScriptRepositoryRawContentUrlPrefix =
-    QStringLiteral("https://raw.githubusercontent.com/qownnotes/scripts/master/");
+    QStringLiteral("https://raw.githubusercontent.com/qownnotes/scripts/main/");
+
+static void captureScriptChangelogHtml(const MD_CHAR *data, MD_SIZE dataSize, void *userData) {
+    if (dataSize > 0) {
+        static_cast<QByteArray *>(userData)->append(data, int(dataSize));
+    }
+}
 
 Script::Script()
     : id{0}, name(QLatin1String("")), scriptPath(QLatin1String("")), priority{0}, enabled{true} {}
@@ -486,8 +494,70 @@ QUrl Script::remoteFileUrl(const QString &fileName) const {
         return QUrl();
     }
 
-    return QUrl(QStringLiteral("https://raw.githubusercontent.com/qownnotes/scripts/master/") +
-                identifier + QStringLiteral("/") + fileName);
+    return QUrl(ScriptRepositoryRawContentUrlPrefix + identifier + QStringLiteral("/") + fileName);
+}
+
+QUrl Script::remoteChangelogUrl() const { return remoteFileUrl(QStringLiteral("CHANGELOG.md")); }
+
+QUrl Script::repositoryChangelogUrl() const {
+    if (identifier.isEmpty()) {
+        return {};
+    }
+
+    return QUrl(QStringLiteral("https://github.com/qownnotes/scripts/blob/main/") + identifier +
+                QStringLiteral("/CHANGELOG.md"));
+}
+
+QString Script::changelogForVersionRange(const QString &changelog, const QString &installedVersion,
+                                         const QString &targetVersion) {
+    const QRegularExpression headingExpression(
+        QStringLiteral("^##[\\t ]+([0-9][^\\t \\r\\n]*)[^\\r\\n]*\\r?$"),
+        QRegularExpression::MultilineOption);
+    QRegularExpressionMatchIterator iterator = headingExpression.globalMatch(changelog);
+    QList<QRegularExpressionMatch> headings;
+
+    while (iterator.hasNext()) {
+        headings.append(iterator.next());
+    }
+
+    const VersionNumber localVersion(installedVersion);
+    const VersionNumber remoteVersion(targetVersion);
+    QStringList sections;
+
+    for (int i = 0; i < headings.count(); ++i) {
+        const QRegularExpressionMatch &heading = headings.at(i);
+        const VersionNumber entryVersion(heading.captured(1));
+
+        if (!(localVersion < entryVersion && entryVersion <= remoteVersion)) {
+            continue;
+        }
+
+        const int start = heading.capturedStart();
+        const int end =
+            i + 1 < headings.count() ? headings.at(i + 1).capturedStart() : changelog.length();
+        sections.append(changelog.mid(start, end - start).trimmed());
+    }
+
+    return sections.join(QStringLiteral("\n\n"));
+}
+
+QString Script::changelogHtmlForVersionRange(const QString &changelog,
+                                             const QString &installedVersion,
+                                             const QString &targetVersion) {
+    const QByteArray markdown =
+        changelogForVersionRange(changelog, installedVersion, targetVersion).toUtf8();
+    if (markdown.isEmpty()) {
+        return {};
+    }
+
+    QByteArray html;
+    if (md_html(markdown.constData(), MD_SIZE(markdown.size()), &captureScriptChangelogHtml, &html,
+                MD_DIALECT_GITHUB, 0) != 0) {
+        qWarning() << "Could not render the script changelog";
+        return {};
+    }
+
+    return QString::fromUtf8(html);
 }
 
 /**